Сделал рефакторинг ImagesListViewController и ProfileViewController изменил архитектуру с MVC на MVP
