+++ ImageFeed/SplashVC/SplashViewController.swift
@@ -99,6 +99,7 @@ extension SplashViewController {
         let tabBarController = UIStoryboard(name: "Main", bundle: .main)
             .instantiateViewController(withIdentifier: "TabBarViewController")
         window.rootViewController = tabBarController
+        
     }
 }
