+++ ImageFeed/TabBarController/TabBarController.swift
@@ -19,6 +19,9 @@ final class TabBarController: UITabBarController {
         let profileViewController = ProfileViewController()
         imagesListViewController.tabBarItem = UITabBarItem(title: "", image: UIImage(named: "tab_editorial_active"), selectedImage: nil)
         profileViewController.tabBarItem = UITabBarItem(title: "", image: UIImage(named: "tab_profile_active"), selectedImage: nil)
+        profileViewController.configure(ProfilePresenter())
+//        profileViewController.presenter = ProfilePresenter()
+//        profileViewController.presenter?.view = profileViewController
         self.viewControllers = [imagesListViewController, profileViewController]
     }
 }
